Refactor to separate timecode parstringifying and React components
